Refactor ONTRun component and related store for creating and updating runs
- Updated the ONTRun component to use the new useOntRunCreateStore for managing run state.
- Removed the old ONTRun.spec.js test file and replaced it with ONTRunShow.vue and its corresponding test file.
- Implemented a new store (ontRunCreate.js) to handle the creation and updating of ONT runs.
- Added comprehensive tests for the new store, ensuring all actions and getters function as expected.
- Updated flow cell handling logic to ensure proper validation and state management.
