+++ src/components/ont/runs/ONTRunInformation.vue
@@ -37,15 +37,15 @@
 
 <script setup>
 import { computed } from 'vue'
-import { useOntRunsStore } from '@/stores/ontRuns'
-import useOntRootStore from '@/stores/ontRoot'
+import { useOntRunCreateStore } from '@/stores/ontRunCreate.js'
+import useOntRootStore from '@/stores/ontRoot.js'
 
 // Initialize stores
-const ontRunsStore = useOntRunsStore()
+const ontRunCreateStore = useOntRunCreateStore()
 const ontRootStore = useOntRootStore()
 
 const { instruments } = ontRootStore
-const { setInstrumentName, setState, setRebasecallingProcess, currentRun } = ontRunsStore
+const { setInstrumentName, setState, setRebasecallingProcess, currentRun } = ontRunCreateStore
 
 // Static lists
 const statesList = ['Pending', 'Completed', 'User Terminated', 'Instrument Crashed', 'Restart']
@@ -82,7 +82,7 @@ const rebasecallingOptions = computed(() => [
   })),
 ])
 
-const newRecord = computed(() => isNaN(ontRunsStore.currentRun.id))
+const newRecord = computed(() => isNaN(ontRunCreateStore.currentRun.id))
 
 // Helper function to format state values
 const formatState = (str) => str.replace(/\s+/g, '_').toLowerCase()
