+++ src/components/ont/runs/ONTRunInstrumentFlowcells.vue
@@ -19,9 +19,9 @@
 
 <script>
 import { mapState } from 'pinia'
-import { useOntRunsStore } from '@/stores/ontRuns'
-import useOntRootStore from '@/stores/ontRoot'
-import ONTFlowcell from '@/components/ont/runs/ONTFlowcell'
+import { useOntRunCreateStore } from '@/stores/ontRunCreate.js'
+import useOntRootStore from '@/stores/ontRoot.js'
+import ONTFlowcell from '@/components/ont/runs/ONTFlowcell.vue'
 /**
  * # ONTRunInstrumentFlowcells
  *
@@ -35,7 +35,7 @@ export default {
   },
   props: {},
   computed: {
-    ...mapState(useOntRunsStore, ['currentRun']),
+    ...mapState(useOntRunCreateStore, ['currentRun']),
     ...mapState(useOntRootStore, ['instruments', 'instrumentByName']),
     instrument() {
       return this.instrumentByName(this.currentRun.instrument_name)
